Remove Ashby job integration from Join us page
